backend/requirements.txt
Django==6.0.3
django-cors-headers==4.9.0
django-mongodb-backend==6.0.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
gunicorn==23.0.0
pymongo==4.16.0
python-dotenv==1.2.1
whitenoise==6.9.0
